+++ Sources/PackageGraph/PinsStore.swift
@@ -12,7 +12,6 @@ import Basic
 import SPMUtility
 import SourceControl
 import PackageModel
-import PackageGraph
 
 public final class PinsStore {
     public struct Pin {
@@ -45,7 +44,7 @@ public final class PinsStore {
     /// The pins map.
     ///
     /// Key -> Package Identity.
-    fileprivate(set) var pinsMap: [String: Pin]
+    public fileprivate(set) var pinsMap: [String: Pin]
 
     /// The current pins.
     public var pins: AnySequence<Pin> {
@@ -71,7 +70,7 @@ public final class PinsStore {
         do {
             _ = try self.persistence.restoreState(self)
         } catch SimplePersistence.Error.restoreFailure {
-            throw WorkspaceDiagnostics.MalformedPackageResolved()
+            throw MalformedPackageResolved()
         }
     }
 
@@ -97,25 +96,6 @@ public final class PinsStore {
     /// This will replace any previous pin with same package name.
     public func add(_ pin: Pin) {
         pinsMap[pin.packageRef.identity] = pin
-    }
-
-    /// Pin a managed dependency at its checkout state.
-    ///
-    /// This method does nothing if the dependency is in edited state.
-    func pin(_ dependency: ManagedDependency) {
-
-        // Get the checkout state.
-        let checkoutState: CheckoutState
-        switch dependency.state {
-        case .checkout(let state):
-            checkoutState = state
-        case .edited, .local:
-            return
-        }
-
-        self.pin(
-            packageRef: dependency.packageRef,
-            state: checkoutState)
     }
 
     /// Unpin all of the currently pinnned dependencies.
@@ -183,47 +163,8 @@ extension PinsStore.Pin: JSONMappable, JSONSerializable, Equatable {
     }
 }
 
-/// A file watcher utility for the Package.resolved file.
-///
-/// This is not intended to be used directly by clients.
-final class ResolvedFileWatcher {
-    private var fswatch: FSWatch!
-    private var existingValue: ByteString?
-    private let valueLock: Lock = Lock()
-    private let resolvedFile: AbsolutePath
-
-    public func updateValue() {
-        valueLock.withLock {
-            self.existingValue = try? localFileSystem.readFileContents(resolvedFile)
-        }
-    }
-
-    init(resolvedFile: AbsolutePath, onChange: @escaping () -> ()) throws {
-        self.resolvedFile = resolvedFile
-
-        let block = { [weak self] (paths: [AbsolutePath]) in
-            guard let self = self else { return }
-
-            // Check if resolved file is part of the received paths.
-            let hasResolvedFile = paths.contains{ $0.appending(component: resolvedFile.basename) == resolvedFile }
-            guard hasResolvedFile else { return }
-
-            self.valueLock.withLock {
-                // Compute the contents of the resolved file and fire the onChange block
-                // if its value is different than existing value.
-                let newValue = try? localFileSystem.readFileContents(resolvedFile)
-                if self.existingValue != newValue {
-                    self.existingValue = newValue
-                    onChange()
-                }
-            }
-        }
-
-        fswatch = FSWatch(paths: [resolvedFile.parentDirectory], latency: 1, block: block)
-        try fswatch.start()
-    }
-
-    deinit {
-        fswatch.stop()
+public struct MalformedPackageResolved: Swift.Error, CustomStringConvertible {
+    public var description: String {
+        return "Package.resolved file is corrupted or malformed; fix or delete the file to continue"
     }
 }
